Install Docker Desktop
I don't love that the RPM isn't signed, so it's pinned by the hash. The hash was taken from the Release Notes, although it was hosted on the same place that the artifact was so idk if that helps at all. 
I don't really know where to put the RPM. Hoping that /tmp works