Revert "Refer to the key in /usr/etc instead of /etc"
This reverts commit 5f3a520eea324d5bfe724df56af7e764c134333f.

The `unlinkat` error was caused by an old cosign.pub file which at the
immutable bit set on it. Because of that, and because the filesystem was
mounted read-only, I couldn't delete that ostree deployment.

The solution was to use a live USB to unlock LUKS, mount the filesystem
directly as writeable, remove the immutable bit, then reboot and run
`ostree admin cleanup`

Since that's resolved now and since it makes more sense to refer to /etc
instead of /usr/etc this is getting reverted.